[138] Created biomes map (#147)
* Removed Growth track and extra daily growth features

* Added enough code adjustments for testing

* Refactored and adjusted code to handle better the terrain and tiles

* Adjusted to make Lake and Tundra special biomes under conditions

* Refactored to handle octaves and height spline

* Added and configured code to create all biomes infinitely

+++ My Forest/Assets/_Main/Scripts/Decorations/DecorationsSources.cs
@@ -4,7 +4,7 @@ namespace MyForest
 {
     public interface IDecorationsServiceSource
     {
-        Decoration CreateDecoration(Transform parent, DecorationData treeData, int height, bool withEntryAnimation);
+        Decoration CreateDecoration(Transform parent, DecorationData treeData, float height, bool withEntryAnimation);
         DecorationData GetRandomDecorationDataForBiome(Biome biome);
     }
